session_pipeline: Addressing issue in wrapper that caused ValueError
When running the top down step, we would run into an error that
said "ValueError: setting an array element with a sequence. The
requested array has an inhomogeneous shape after 2 dimensions.
The detected shape was (3352, 133) + inhomogeneous part". If the
bbox is None, I changed it to set the keypoints to 0s that is
2 columns rather than 3. In a future commit I will update this to be
more robust in case the keypoints shape has 3 columns.