top_down: Updating mmpose topdown to include confidence
Getting the keypoint scores and visibility in the mmpose
wrapper. Returning scores normalized by the max value packaged
with the keypoints and additionally returning the raw scores
and visibiliy values. These will eventually be inserted into
the part table that is currently commented out in TopDownPerson.
This part table will be uncommented in a future commit.